+++ src/components/Order/Order.scss
@@ -6,7 +6,7 @@ div.order-sidebar {
   min-height: 100vh;
 }
 
-ul .removeStyle {
+.removeStyle {
   list-style-type: none;
   padding: 1rem;
 }
