+++ src/components/Products/ProductData.tsx
@@ -1,0 +1,110 @@
+const allProducts: AllProducts = [
+  {
+    categoryName: 'Burger',
+    list: [
+      {
+        id: '101',
+        name: 'Chicken burger',
+        image: '/images/chicken.jpg',
+        price: '12',
+        quantity: 0,
+      },
+      {
+        id: '102',
+        name: 'Bacon burger',
+        image: '/images/bacon and cheese.jpg',
+        price: '12',
+        quantity: 0,
+      },
+      {
+        id: '103',
+        name: 'Vegan burger',
+        image: '/images/vegan.jpg',
+        price: '12',
+        quantity: 0,
+      },
+      {
+        id: '104',
+        name: 'Beef burger',
+        image: '/images/beef.jpg',
+        price: '12',
+        quantity: 0,
+      },
+      {
+        id: '105',
+        name: 'Veggie burger',
+        image: '/images/veggie.jpg',
+        price: '12',
+        quantity: 0,
+      },
+      {
+        id: '106',
+        name: 'Lamb burger',
+        image: '/images/lamb.jpg',
+        price: '12',
+        quantity: 0,
+      },
+    ],
+  },
+  {
+    categoryName: 'Sides',
+    list: [
+      {
+        id: '201',
+        name: 'Veggie Salad',
+        image: '/images/caesar.jpg',
+        price: '7',
+        quantity: 0,
+      },
+      {
+        id: '202',
+        name: 'Caeser Salad',
+        image: '/images/salad.jpg',
+        price: '7',
+        quantity: 0,
+      },
+      {
+        id: '203',
+        name: 'Small Fries',
+        image: '/images/small fries.jpg',
+        price: '7',
+        quantity: 0,
+      },
+      {
+        id: '204',
+        name: 'Large Fries',
+        image: '/images/largefries.jpg',
+        price: '7',
+        quantity: 0,
+      },
+    ],
+  },
+  {
+    categoryName: 'Drinks',
+    list: [
+      {
+        id: '301',
+        name: 'Coke',
+        image: '/images/pepsi.jpg',
+        price: '4',
+        quantity: 0,
+      },
+      {
+        id: '302',
+        name: 'Kombucha',
+        image: '/images/kombuch.jpg',
+        price: '4',
+        quantity: 0,
+      },
+      {
+        id: '303',
+        name: 'Orange Juice',
+        image: '/images/orange juice.jpg',
+        price: '4',
+        quantity: 0,
+      },
+    ],
+  },
+];
+
+export default allProducts;
