+++ src/components/Products/Products.tsx
@@ -5,11 +5,12 @@ import Checkout from '../Checkout';
 import Order from '../Order';
 import './Products.scss';
 import ProductList from '../ProductList';
+import allProducts from './ProductData';
 import { transactionFlow as transactionFlowService } from '../../services';
 
 function productClick(
   shortlistedProducts: Array<Product>,
-  allProducts: AllProducts,
+  productList: AllProducts,
   id: string,
   updateShortlistedProducts: Function
 ) {
@@ -29,7 +30,7 @@ function productClick(
     // else find the clicked product details in allProducts array and insert it at top
     // of short listed product list.
     let clickedProduct: Product;
-    allProducts.forEach((c: CategoryProducts) =>
+    productList.forEach((c: CategoryProducts) =>
       c.list.forEach((p: Product) => {
         if (p.id === id) {
           clickedProduct = p;
@@ -139,115 +140,6 @@ function Products({
   openPricing,
   setOpenPricing,
 }: Props) {
-  const allProducts: AllProducts = [
-    {
-      categoryName: 'Burger',
-      list: [
-        {
-          id: '101',
-          name: 'Chicken burger',
-          image: '/images/chicken.jpg',
-          price: '12',
-          quantity: 0,
-        },
-        {
-          id: '102',
-          name: 'Bacon burger',
-          image: '/images/bacon and cheese.jpg',
-          price: '12',
-          quantity: 0,
-        },
-        {
-          id: '103',
-          name: 'Vegan burger',
-          image: '/images/vegan.jpg',
-          price: '12',
-          quantity: 0,
-        },
-        {
-          id: '104',
-          name: 'Beef burger',
-          image: '/images/beef.jpg',
-          price: '12',
-          quantity: 0,
-        },
-        {
-          id: '105',
-          name: 'Veggie burger',
-          image: '/images/veggie.jpg',
-          price: '12',
-          quantity: 0,
-        },
-        {
-          id: '106',
-          name: 'Lamb burger',
-          image: '/images/lamb.jpg',
-          price: '12',
-          quantity: 0,
-        },
-      ],
-    },
-    {
-      categoryName: 'Sides',
-      list: [
-        {
-          id: '201',
-          name: 'Veggie Salad',
-          image: '/images/caesar.jpg',
-          price: '7',
-          quantity: 0,
-        },
-        {
-          id: '202',
-          name: 'Caeser Salad',
-          image: '/images/salad.jpg',
-          price: '7',
-          quantity: 0,
-        },
-        {
-          id: '203',
-          name: 'Small Fries',
-          image: '/images/small fries.jpg',
-          price: '7',
-          quantity: 0,
-        },
-        {
-          id: '204',
-          name: 'Large Fries',
-          image: '/images/largefries.jpg',
-          price: '7',
-          quantity: 0,
-        },
-      ],
-    },
-    {
-      categoryName: 'Drinks',
-      list: [
-        {
-          id: '301',
-          name: 'Coke',
-          image: '/images/pepsi.jpg',
-          price: '4',
-          quantity: 0,
-        },
-        {
-          id: '302',
-          name: 'Kombucha',
-          image: '/images/kombuch.jpg',
-          price: '4',
-          quantity: 0,
-        },
-        {
-          id: '303',
-          name: 'Orange Juice',
-          image: '/images/orange juice.jpg',
-          price: '4',
-          quantity: 0,
-        },
-      ],
-    },
-  ];
-
   const [shortlistedProducts, updateShortlistedProducts] = useState<Array<Product>>([]);
   const [checkout, setCheckout] = useState(false);
   const [transactionAction, setTransactionAction] = useState('');
