+++ src/components/RefundPay/RefundPay.test.tsx
@@ -12,21 +12,21 @@ describe('RefundPay', () => {
   });
 
   it('should return 0 when refund button clicked without entering value', () => {
-    component.find('button').simulate('click');
+    component.find('button#refundButton').simulate('click');
     expect(handleRefundPay.mock.calls.length).toBe(1);
     expect(handleRefundPay.mock.calls[0][0]).toBe(0);
   });
 
   it('should return $1 when refund button clicked after entering value 100 cents', () => {
-    component.find('#Refund-Amount').simulate('change', { target: { value: 100 } });
+    component.find('#inpRefundAmount').simulate('change', { target: { value: 100 } });
 
-    component.find('button').simulate('click');
+    component.find('button#refundButton').simulate('click');
     expect(handleRefundPay.mock.calls[0][0]).toBe(1);
   });
 
   it('should return $1 when refund button clicked after entering value 100 cents', () => {
-    component.find('#Refund-Amount').simulate('change', { target: { value: 100 } });
-    component.find('button').simulate('click');
+    component.find('#inpRefundAmount').simulate('change', { target: { value: 100 } });
+    component.find('button#refundButton').simulate('click');
     expect(handleRefundPay.mock.calls[0][0]).toBe(1);
   });
 });
